Type to store function that acts as linear ops
This type can be used to create arrays of operators (+transpose, +proximals) that are really just
std::function<> objects. There is an accompanying wrapper to transform an matrix into a pair of
functions. The type federates the behavior of functions and matrices into a single object, called
via the * operator.