Weights must be positive
The weights associated with the l1 proximal in the regression example
where negative in some cases (generated as random number).
The c and cpp examples both run in that case, but with some variations
in the implementation (projection on the positive quadrant?).
In pratices, a soft-threshhold function can be defined with negative
values, so catching this error has happen on the user side, e.g. by
including defensive programming stuff.